+++ lib/vmpooler/providers/dummy.rb
@@ -1,4 +1,5 @@
 require 'yaml'
+require 'vmpooler/providers/base'
 
 module Vmpooler
   class PoolManager